Make `const_fn` an optional dependency
The docs for the latest `const_fn` have a snippet on how to use the
attribute macro as an optional dependency [0].

[0] https://docs.rs/const_fn/0.4.9/const_fn/#use-this-crate-as-an-optional-dependency

These two attributes have the same effect:

    #[const_fn(feature = "const")]
    pub const fn whatever() {}

    #[cfg_attr(feature = "const", const_fn::const_fn)]
    pub fn whatever() {}

But the `cfg_attr` version only requires `const_fn` crate to be a
dependency when the `const` feature is enabled.

This enables building `tz-rs` without any dependencies and without any
proc macros.

This was mostly an automated refactor with `sed`. I took a peek at the
docs with `cargo doc --open` to see if things looked mostly the same
with default features.